+++ src/neuroconv/datainterfaces/__init__.py
@@ -1,6 +1,7 @@
 # Behavior
 from .behavior.audio.audiointerface import AudioInterface
 from .behavior.dannce.danncedatainterface import DANNCEInterface
+from .behavior.dannce.sdanncedatainterface import SDANNCEInterface
 from .behavior.deeplabcut.deeplabcutdatainterface import DeepLabCutInterface
 from .behavior.fictrac.fictracdatainterface import FicTracDataInterface
 from .behavior.lightningpose.lightningposedatainterface import (
@@ -186,6 +187,7 @@ interface_list = [
     InternalVideoInterface,
     AudioInterface,
     DANNCEInterface,
+    SDANNCEInterface,
     DeepLabCutInterface,
     SLEAPInterface,
     MiniscopeBehaviorInterface,
@@ -234,6 +236,7 @@ interfaces_by_category = dict(
         ExternalVideo=ExternalVideoInterface,
         InternalVideo=InternalVideoInterface,
         DANNCE=DANNCEInterface,
+        SDANNCE=SDANNCEInterface,
         DeepLabCut=DeepLabCutInterface,
         SLEAP=SLEAPInterface,
         FicTrac=FicTracDataInterface,